chore(tangle-dapp): Move form to separate compopnent
